moved bout scripts into bout folder and checked working pipeline
