--- src/bout_plotting.py
@@ -1,361 +1,0 @@
-import pandas as pd
-import matplotlib.pyplot as plt
-import matplotlib.patches as patches
-import numpy as np
-
-import scipy.stats as stats
-import bout_clustering as bt_clustering
-
-from core import FREQUENCY_COLOR_MAPPINGS
-
-def plot_ipi_hist(location_df, fig_details):
-    """
-    Plots a histogram of IPIs from a provided location and frequency group for the specified range of intervals and bin width.
-    """
-
-    ipis_ms = bt_clustering.get_valid_ipis_ms(location_df)
-
-    plt.rcParams.update({'font.size' : 20})
-    plt.figure(figsize=(12, 6))
-    plt.title(f"Frequency histogram for {fig_details['freq_group']}inter-pulse intervals {fig_details['site_name']} (bin size = {fig_details['bin_step']})", loc='left', fontsize=20)
-    hist_bins = np.arange(0, fig_details['time_end']+fig_details['bin_step'], fig_details['bin_step'])
-    histogram = plt.hist(ipis_ms, bins=hist_bins, facecolor='y', edgecolor='k')
-    plt.grid(axis='y')
-    bin_ticks = hist_bins.round(3)
-    bin_labels = [''] * len(bin_ticks)
-
-    bin_labels[::fig_details['time_end']//(10*fig_details['bin_step'])] = bin_ticks[::fig_details['time_end']//(10*fig_details['bin_step'])]
-
-    plt.xticks(bin_ticks, bin_labels, rotation=45)
-    plt.ylabel('# of intervals', fontsize=20)
-    plt.xlabel('Time (ms)', loc='left', fontsize=20)
-    plt.xlim(0, fig_details['time_end'])
-    plt.tight_layout()
-    plt.show()
-
-    return ipis_ms, histogram
-
-def plot_log_survival(intervals_ms, survival, fig_details):
-    """
-    Plots the log-survivorship curve of intervals with 10ms resolution with x-axis in minutes.
-    """
-
-    intervals_s = np.divide(intervals_ms, 1000).astype('int')
-    intervals_min = np.divide(intervals_s, 60).astype('int')
-
-    plt.figure(figsize=(12, 6))
-    plt.rcParams.update({'font.size' : 20})
-    plt.title(f"{fig_details['freq_group']}log-survivorship curve in {fig_details['site_name']}", fontsize=20)
-    plt.plot(intervals_ms, survival, marker='.', c='navy')
-    plt.ylabel('log(# of intervals)', fontsize=18)
-    plt.xlabel('Time (ms)', loc='left', fontsize=18)
-    plt.xticks(intervals_ms[::180000], intervals_ms[::180000], rotation=45)
-    plt.grid(which='both')
-    plt.xlim(10, 10e6)
-    plt.xscale("log")
-
-    plt.tight_layout()
-    plt.show()
-
-def plot_fast_regression_results(intervals_ms, survival, fast_process, fig_details):
-    """
-    Plots the log-survivorship curve of intervals in the time window to visualize the linear regression of the fast process
-    """
-
-    fast_start = (intervals_ms[fast_process['indices']][0])
-    fast_end = (intervals_ms[fast_process['indices']][-1])
-
-    print(f'Considering points from {fast_start} to {fast_end} for fast process')
-    plt.figure(figsize=(12, 6))
-    plt.rcParams.update({'font.size' : 20})
-    plt.title(f"{fig_details['freq_group']}log-survivorship curve in {fig_details['site_name']}", fontsize=20)
-
-    plt.plot(intervals_ms, survival, marker='.', c='cyan', alpha=0.8)
-    plt.plot(intervals_ms[fast_process['indices']], survival[fast_process['indices']], marker='.', c='navy')
-    plt.axvline(fast_start, linestyle='dashed', color='k')
-    plt.axvline(fast_end, linestyle='dashed', color='k')
-
-    x = np.linspace(-100, fig_details['time_end'])
-    plt.plot(x, fast_process['metrics'].intercept + fast_process['metrics'].slope*x, c='red', marker=None, 
-             label=f"R^2 = {round(fast_process['metrics'].rvalue**2, 4)}, Intercept = {round(fast_process['metrics'].intercept, 2)}, Slope = {round(fast_process['metrics'].slope, 8)}")
-
-    plt.grid(axis='y')
-    plt.ylabel('log(# of intervals)', fontsize=18)
-    plt.xlabel('Time (ms)', fontsize=18)
-    plt.xlim(intervals_ms[0]-100, fig_details['time_end'])
-    plt.ylim(survival[-1]-0.2, survival[0]+0.2)
-    plt.tight_layout()
-    plt.legend(loc='upper right')
-
-    plt.show()
-
-def plot_slow_regression_results(intervals_ms, survival, slow_process, fig_details):
-    """
-    Plots the log-survivorship curve of intervals in the time window to visualize the linear regression of the slow process
-    """
-
-    slow_start = (intervals_ms[slow_process['indices']][0])
-    slow_end = (intervals_ms[slow_process['indices']][-1])
-
-    print(f'Considering points from {slow_start} to {slow_end} for slow process')
-    plt.figure(figsize=(12, 6))
-    plt.rcParams.update({'font.size' : 20})
-    plt.title(f"{fig_details['freq_group']}log-survivorship curve in {fig_details['site_name']}", fontsize=20)
-
-    plt.plot(intervals_ms, survival, marker='.', c='cyan', alpha=0.8)
-    plt.plot(intervals_ms[slow_process['indices']], survival[slow_process['indices']], marker='.', c='navy')
-    plt.axvline(slow_start, linestyle='dashed', color='k')
-    plt.axvline(slow_end, linestyle='dashed', color='k')
-
-    x = np.linspace(-100, intervals_ms[-1], int(slow_end))
-    plt.plot(x, slow_process['metrics'].intercept + slow_process['metrics'].slope*x, c='blue', marker=None, alpha=0.8, 
-             label=f"R^2 = {round(slow_process['metrics'].rvalue**2, 4)}, Intercept = {round(slow_process['metrics'].intercept, 2)}, Slope = {round(slow_process['metrics'].slope, 8)}")
-
-    plt.grid(axis='y')
-    plt.ylabel('log(# of intervals)', fontsize=18)
-    plt.xlabel('Time (ms)', fontsize=18)
-    plt.xlim(intervals_ms[0], intervals_ms[-1])
-    plt.ylim(survival[-1]-0.2, survival[0]+0.2)
-    plt.tight_layout()
-    plt.legend(loc='upper right')
-    
-    plt.show()
-
-def plot_bci_over_survival(intervals_ms, survival, fast_process, slow_process, bci, fig_details):
-    """
-    Plots the log-survivorship curve of intervals in the time window to visualize the linear regression of the both processs
-    along with the derived BCI to see where it is in time.
-    """
-
-    fast_start = (intervals_ms[fast_process['indices']][0])
-    fast_end = (intervals_ms[fast_process['indices']][-1])
-    slow_start = (intervals_ms[slow_process['indices']][0])
-    slow_end = (intervals_ms[slow_process['indices']][-1])
-
-    plt.figure(figsize=(12, 6))
-    plt.rcParams.update({'font.size' : 20})
-    plt.title(f"{fig_details['freq_group']}log-survivorship curve in {fig_details['site_name']}", fontsize=20)
-    plt.plot(intervals_ms, survival, marker='.', c='cyan', alpha=0.8)
-
-    plt.plot(intervals_ms[fast_process['indices']], survival[fast_process['indices']], marker='.', c='navy')
-    plt.plot(intervals_ms[slow_process['indices']], survival[slow_process['indices']], marker='.', c='navy')
-    plt.axvline(fast_start, linestyle='dashed', color='k')
-    plt.axvline(fast_end, linestyle='dashed', color='k')
-    plt.axvline(slow_start, linestyle='dashed', color='k')
-    plt.axvline(slow_end, linestyle='dashed', color='k')
-
-    x = np.linspace(-100, intervals_ms[-1], int(slow_end))
-    plt.plot(x, fast_process['metrics'].intercept + fast_process['metrics'].slope*x, c='red', alpha=0.6, marker=None, label=f"R^2 = {round(fast_process['metrics'].rvalue**2, 4)}")
-    plt.plot(x, slow_process['metrics'].intercept + slow_process['metrics'].slope*x, c='blue', alpha=0.6, marker=None, label=f"R^2 = {round(slow_process['metrics'].rvalue**2, 4)}")
-
-    plt.axvline(bci, linestyle='dashed', linewidth=2, color='yellowgreen', label=f"BCI: {round(bci/1000, 2)}s")
-    plt.grid(axis='y')
-    plt.ylabel('log(# of intervals)', fontsize=18)
-    plt.xlabel('Time (ms)', fontsize=18)
-    plt.xlim(10e0, 10e7)
-    plt.ylim(survival[-1]-0.2, survival[0]+0.2)
-    plt.tight_layout()
-    plt.legend(loc='upper right')
-    plt.xscale("log")
-    plt.show()
-
-def plot_nlin_results_over_survival(intervals_ms, survival, fast_process, slow_process, optim_x, optim_bci, bci, fig_details):
-    """
-    Plots the log-survivorship curve of intervals in the time window to visualize the linear regression of the both processs
-    along with two derived BCIs (Slater BCI and Sibly BCI) to see where they are in time.
-    """
-
-    fast_start = (intervals_ms[fast_process['indices']][0])
-    fast_end = (intervals_ms[fast_process['indices']][-1])
-    slow_start = (intervals_ms[slow_process['indices']][0])
-    slow_end = (intervals_ms[slow_process['indices']][-1])
-
-    plt.figure(figsize=(12, 6))
-    plt.rcParams.update({'font.size' : 20})
-    plt.title(f"{fig_details['freq_group']}log-survivorship curve in {fig_details['site_name']}", fontsize=20)
-    plt.plot(intervals_ms, survival, marker='.', c='cyan', alpha=0.8)
-
-    plt.plot(intervals_ms, bt_clustering.model(intervals_ms, optim_x[0], optim_x[1], optim_x[2], optim_x[3]), marker='.', c='green', alpha=0.1)
-    plt.plot(intervals_ms[fast_process['indices']], survival[fast_process['indices']], marker='.', c='navy')
-    plt.plot(intervals_ms[slow_process['indices']], survival[slow_process['indices']], marker='.', c='navy')
-    plt.axvline(fast_start, linestyle='dashed', color='k')
-    plt.axvline(fast_end, linestyle='dashed', color='k')
-    plt.axvline(slow_start, linestyle='dashed', color='k')
-    plt.axvline(slow_end, linestyle='dashed', color='k')
-
-    x = np.linspace(-100, intervals_ms[-1], int(slow_end))
-    plt.plot(x, fast_process['metrics'].intercept + fast_process['metrics'].slope*x, c='red', alpha=0.6, marker=None, label=f"R^2 = {round(fast_process['metrics'].rvalue**2, 4)}")
-    plt.plot(x, slow_process['metrics'].intercept + slow_process['metrics'].slope*x, c='blue', alpha=0.6, marker=None, label=f"R^2 = {round(slow_process['metrics'].rvalue**2, 4)}")
-
-    plt.axvline(bci, linestyle='dashed', linewidth=2, color='yellowgreen', label=f"BCI: {round(bci/1000, 2)}s")
-    plt.axvline(optim_bci, linestyle='dashed', linewidth=2, color='yellowgreen', label=f"Optim BCI: {round(optim_bci/1000, 2)}s")
-    plt.grid(which='both')
-    plt.ylabel('log(# of intervals)', fontsize=18)
-    plt.xlabel('Time (ms)', fontsize=18)
-    plt.xlim(10e0, 10e6)
-    plt.ylim(survival[-1]-0.2, survival[0]+0.2)
-    plt.tight_layout()
-    plt.legend(loc='upper right')
-    plt.xscale("log")
-    plt.show()
-
-def plot_audio_seg(audio_features, spec_features):
-    """
-    Function to plot the spectrogram of a provided audio segment
-    """
-
-    audio_seg = audio_features['audio_seg']
-    fs = audio_features['sample_rate']
-    start = audio_features['start']
-    duration = audio_features['duration']
-
-    plt.figure(figsize=(15, 5))
-    plt.rcParams.update({'font.size' : 24})
-    plt.title(f"Spectrogram representation of {audio_features['file_path'].name}", fontsize=22)
-    plt.rcParams.update({'font.size': 24})
-    plt.specgram(audio_seg, NFFT=spec_features['NFFT'], cmap=spec_features['cmap'], vmin=spec_features['vmin'])
-    plt.yticks(ticks=np.linspace(0, 1, 6), labels=np.linspace(0, fs/2000, 6).astype('int'))
-    plt.xticks(ticks=np.linspace(0, duration*(fs/2), 11), labels=np.round(np.linspace(start, start+duration, 11, dtype='float'), 2), rotation=30)
-    plt.ylabel("Frequency (kHz)")
-    plt.xlabel("Time (s)")
-    plt.gcf().autofmt_xdate()
-
-    plt.tight_layout()
-    plt.show()
-
-def plot_dets_over_audio_seg(audio_features, spec_features, plot_dets):
-    """
-    Function to plot the spectrogram of a provided audio segment with overlayed detections
-    """
-
-    audio_seg = audio_features['audio_seg']
-    fs = audio_features['sample_rate']
-    start = audio_features['start']
-    duration = audio_features['duration']
-
-    plt.figure(figsize=(15, 5))
-    plt.title(f"BatDetect2 detections on {audio_features['file_path'].name}", fontsize=22)
-    plt.rcParams.update({'font.size': 24})
-    plt.specgram(audio_seg, NFFT=spec_features['NFFT'], cmap=spec_features['cmap'], vmin=spec_features['vmin'])
-
-    yellow_patch = patches.Patch(facecolor='yellow', edgecolor='k', label='Detections')
-
-    legend_patches = [yellow_patch]
-    ax = plt.gca()
-    for i, row in plot_dets.iterrows():
-        rect = patches.Rectangle(((row['start_time'] - start)*(fs/2), row['low_freq']/(fs/2)), 
-                        (row['end_time'] - row['start_time'])*(fs/2), (row['high_freq'] - row['low_freq'])/(fs/2), 
-                        linewidth=2, edgecolor='yellow', facecolor='none', alpha=0.8)
-        
-        ax.add_patch(rect)
-
-    plt.yticks(ticks=np.linspace(0, 1, 6), labels=np.linspace(0, fs/2000, 6).astype('int'))
-    plt.xticks(ticks=np.linspace(0, duration*(fs/2), 11), labels=np.round(np.linspace(start, start+duration, 11, dtype='float'), 2), rotation=30)
-    plt.ylabel("Frequency (kHz)")
-    plt.xlabel("Time (s)")
-    plt.gcf().autofmt_xdate()
-    plt.legend(handles=legend_patches, fontsize=20, ncol=int(len(legend_patches)**0.5), loc='upper right')
-
-    plt.tight_layout()
-    plt.show()
-
-def plot_dets_with_bout_ID_over_audio_seg(audio_features, spec_features, data_params, plot_dets):
-    """
-    Function to plot the spectrogram of a provided audio segment with overlayed detections and bout tags for bout starts and ends.
-    If there is a duty-cycle scheme being applied, also plots the recording windows being simulated.
-    """
-
-    audio_seg = audio_features['audio_seg']
-    fs = audio_features['sample_rate']
-    start = audio_features['start']
-    duration = audio_features['duration']
-
-    plt.figure(figsize=(15, 5))
-    plt.title(f"BatDetect2 detections for simulated recording period on {audio_features['file_path'].name}", fontsize=22)
-    plt.rcParams.update({'font.size': 24})
-    plt.specgram(audio_seg, NFFT=spec_features['NFFT'], cmap=spec_features['cmap'], vmin=spec_features['vmin'])
-
-    pink_patch = patches.Patch(facecolor='pink', edgecolor='k', label='Bout START/END')
-    yellow_patch = patches.Patch(facecolor='yellow', edgecolor='k', label='Detections')
-
-    legend_patches = [pink_patch, yellow_patch]
-    ax = plt.gca()
-    for i, row in plot_dets.iterrows():
-        if row['call_status'] != 'within bout' and row['call_status'] != 'outside bout':
-            rect = patches.Rectangle(((row['start_time'] - start - 0.01)*(fs/2), (row['low_freq']-2000)/(fs/2)), 
-                        0.04*(fs/2), 12000/(fs/2), linewidth=4, edgecolor='pink', facecolor='none', alpha=0.8)
-        else:
-            rect = patches.Rectangle(((row['start_time'] - start)*(fs/2), row['low_freq']/(fs/2)), 
-                            (row['end_time'] - row['start_time'])*(fs/2), (row['high_freq'] - row['low_freq'])/(fs/2), 
-                            linewidth=2, edgecolor='yellow', facecolor='none', alpha=0.8)
-        
-        ax.add_patch(rect)
-
-    dc_tag = data_params['cur_dc_tag']
-    cycle_length = int(dc_tag.split('of')[1])
-    time_on = int(dc_tag.split('of')[0])
-
-    window_starts = np.arange(0, 1800, cycle_length)
-    for window_start in window_starts:
-        if (window_start <= start+duration and window_start+time_on > start):
-            if (window_start <= start):
-                rect = patches.Rectangle((0, 0), (min(time_on, ((window_start+time_on)-start)))*fs/2, fs/2, 
-                                        linewidth=4, edgecolor='yellow', facecolor='yellow', alpha=0.3)
-            elif (window_start+time_on) >= (start+duration):
-                rect = patches.Rectangle(((window_start-start)*fs/2, 0), 
-                                        (min(time_on, (start+duration) - window_start))*fs/2, fs/2, 
-                                        linewidth=4, edgecolor='yellow', facecolor='yellow', alpha=0.3)
-            else:
-                rect = patches.Rectangle(((window_start-start)*fs/2, 0), 
-                                            (time_on)*fs/2, fs/2, 
-                                            linewidth=4, edgecolor='yellow', facecolor='yellow', alpha=0.3)
-            ax.add_patch(rect)
-
-    plt.yticks(ticks=np.linspace(0, 1, 6), labels=np.linspace(0, fs/2000, 6).astype('int'))
-    plt.xticks(ticks=np.linspace(0, duration*(fs/2), 11), labels=np.round(np.linspace(start, start+duration, 11, dtype='float'), 2), rotation=30)
-    plt.ylabel("Frequency (kHz)")
-    plt.xlabel("Time (s)")
-    plt.gcf().autofmt_xdate()
-    plt.legend(handles=legend_patches, fontsize=20, ncol=int(len(legend_patches)**0.5), loc='upper right')
-
-    plt.tight_layout()
-    plt.show()
-
-def plot_bouts_over_audio_seg(audio_features, spec_features, bout_params, plot_bouts):
-    """
-    Function to plot the spectrogram of a provided audio segment with overlayed bouts and bout duurations.
-    """
-
-    audio_seg = audio_features['audio_seg']
-    fs = audio_features['sample_rate']
-    start = audio_features['start']
-    duration = audio_features['duration']
-
-    plt.figure(figsize=(15, 5))
-    plt.title(f"BCI-derived bouts using BD2 detections on {audio_features['file_path'].name}", fontsize=22)
-    plt.rcParams.update({'font.size': 24})
-    plt.specgram(audio_seg, NFFT=spec_features['NFFT'], cmap=spec_features['cmap'], vmin=spec_features['vmin'])
-
-    yellow_patch = patches.Patch(facecolor='yellow', edgecolor='k', label=f'BCI = {round(bout_params["HF2_bci"], 2)}ms')
-    red_patch = patches.Patch(facecolor='red', edgecolor='k', label=f'BCI = {round(bout_params["HF1_bci"], 2)}ms')
-    blue_patch = patches.Patch(facecolor='cyan', edgecolor='k', label=f'BCI = {round(bout_params["LF1_bci"], 2)}ms')
-
-    legend_patches = [blue_patch, red_patch, yellow_patch]
-    ax = plt.gca()
-    for i, row in plot_bouts.iterrows():
-        plt.text(x=(row['start_time'] - start + (row['bout_duration_in_secs']/5))*(fs/2), y=min((row['high_freq']+2000)/(fs/2), 3/4), 
-                            s=f"{round(row['bout_duration_in_secs'], 2)}s", color='pink', fontsize=14)
-        rect = patches.Rectangle(((row['start_time'] - start)*(fs/2), row['low_freq']/(fs/2)), 
-                        (row['bout_duration_in_secs'])*(fs/2), (row['high_freq'] - row['low_freq'])/(fs/2), 
-                        linewidth=2, edgecolor=FREQUENCY_COLOR_MAPPINGS[row['freq_group']], facecolor='none', alpha=0.8)
-        ax.add_patch(rect)
-
-    plt.yticks(ticks=np.linspace(0, 1, 6), labels=np.linspace(0, fs/2000, 6).astype('int'))
-    plt.xticks(ticks=np.linspace(0, duration*(fs/2), 11), labels=np.round(np.linspace(start, start+duration, 11, dtype='float'), 2), rotation=30)
-    plt.ylabel("Frequency (kHz)")
-    plt.xlabel("Time (s)")
-    plt.gcf().autofmt_xdate()
-    plt.legend(handles=legend_patches, fontsize=20, ncol=int(len(legend_patches)), loc='upper right')
-
-    plt.tight_layout()
-    plt.show()
